feat(web): refactor theme switching and update dependencies
Introduced a new theme-switch route and refactored the theme
switching logic into a dedicated component. This modularizes the
code, improving maintainability. Also, the commit updates several
dependencies within the `pnpm-lock.yaml` and `package.json` files to
align with the project's requirements.
