+++ apps/web/package.json
@@ -5,6 +5,7 @@
 		}
 	},
 	"dependencies": {
+		"@conform-to/react": "1.2.2",
 		"@epic-web/client-hints": "1.3.5",
 		"@epic-web/invariant": "1.0.0",
 		"@octokit/core": "6.1.2",
@@ -21,7 +22,8 @@
 		"isbot": "5.1.17",
 		"react": "18.3.1",
 		"react-dom": "18.3.1",
-		"react-router": "7.0.2"
+		"react-router": "7.0.2",
+		"remix-utils": "7.7.0"
 	},
 	"devDependencies": {
 		"@codecov/vite-plugin": "1.5.1",
